third_party/blink/web_tests/external/wpt/navigation-api/ordering-and-transition/navigate-204-205-download-then-same-document-expected.txt

This is a testharness.js-based test.
[FAIL] event and promise ordering when navigate() is to a 204s and then to a same-document navigation
  assert_array_equals: expected property 7 to be "committed fulfilled 2" but got "navigatesuccess" (expected array ["navigate", "AbortSignal abort", "navigateerror", "navigate", "currententrychange", "committed rejected 1", "finished rejected 1", "committed fulfilled 2", "promise microtask", "navigatesuccess", "finished fulfilled 2"] got ["navigate", "AbortSignal abort", "navigateerror", "navigate", "currententrychange", "committed rejected 1", "finished rejected 1", "navigatesuccess", "committed fulfilled 2", "promise microtask", "finished fulfilled 2"])
[FAIL] event and promise ordering when navigate() is to a 205s and then to a same-document navigation
  assert_array_equals: expected property 7 to be "committed fulfilled 2" but got "navigatesuccess" (expected array ["navigate", "AbortSignal abort", "navigateerror", "navigate", "currententrychange", "committed rejected 1", "finished rejected 1", "committed fulfilled 2", "promise microtask", "navigatesuccess", "finished fulfilled 2"] got ["navigate", "AbortSignal abort", "navigateerror", "navigate", "currententrychange", "committed rejected 1", "finished rejected 1", "navigatesuccess", "committed fulfilled 2", "promise microtask", "finished fulfilled 2"])
[FAIL] event and promise ordering when navigate() is to a Content-Disposition: attachment responses and then to a same-document navigation
  assert_array_equals: expected property 7 to be "committed fulfilled 2" but got "navigatesuccess" (expected array ["navigate", "AbortSignal abort", "navigateerror", "navigate", "currententrychange", "committed rejected 1", "finished rejected 1", "committed fulfilled 2", "promise microtask", "navigatesuccess", "finished fulfilled 2"] got ["navigate", "AbortSignal abort", "navigateerror", "navigate", "currententrychange", "committed rejected 1", "finished rejected 1", "navigatesuccess", "committed fulfilled 2", "promise microtask", "finished fulfilled 2"])
Harness: the test ran to completion.
